requirements.txt
openai>=1.30.0
pandas>=2.1.0
SQLAlchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.6.0
tenacity>=8.2.0
tqdm>=4.66.0
rich>=13.7.0
PyYAML>=6.0.1
Flask>=3.0.0
